Move SysML test setup to its own fixtures module
