+++ gaphor/SysML/tests/conftest.py
@@ -1,1 +1,1 @@
-from gaphor.diagram.tests.fixtures import diagram, element_factory, event_manager
+from gaphor.SysML.tests.fixtures import diagram, element_factory, event_manager
